Fix device group permissions
Forcing the group permission bits to 0 completely disallows access to
the device by members of the owner group even if the device is allowed
to be accessed by users from other groups. This is too restrictive. The
goal was to ensure the owner group gets the same permission as users
from other groups.

This copies the permission bits from 'other' to 'group' when creating
devices.